Split recipe
